pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "html2text>=2025.4.15",
    "ipython>=9.4.0",
    "jupyter>=1.1.1",
    "markdown>=3.8.2",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "openai>=1.97.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
]
